[ci skip] include _alt_ stuff so combo of Rcpp and package registration works.

+++ tools/install_quick.sh
@@ -7,8 +7,10 @@ set -x
 icd_install_tmp=$(mktemp -d)
 
 function finish {
-        rm -rf "$icd_install_tmp"
-        popd
+  if [ ${1:-} -eq "--clean" ]; then
+    rm -rf "$icd_install_tmp"
+  fi
+  popd
 }
 trap finish EXIT
 
@@ -16,4 +18,4 @@ cp -r "${ICD_HOME:-$HOME/rprojects/icd}" "$icd_install_tmp"
 pushd "$icd_install_tmp"
 mkdir "install"
 R CMD build --no-build-vignettes --no-resave-data icd
-R CMD INSTALL -d --library="install" --install-tests --no-docs "$(ls -t icd*.tar.gz | head -1)"
+R CMD INSTALL --no-clean-on-error -d --library="install" --install-tests --no-docs "$(ls -t icd*.tar.gz | head -1)"
